+++ src/graphics/SpriteFont.cpp
@@ -309,7 +309,7 @@ void vg::SpriteFont::draw(SpriteBatch* batch, const cString s, const f32v2& posi
     if (pos.x < clipRect.x) pos.x = clipRect.x;
     f32 gx = 0.0f;
     std::vector <std::vector<GlyphToRender> > rows(1); // Rows of glyphs
-    std::vector <int> rightEdges(1, 0.0f); // Right edge positions for rows
+    std::vector <f32> rightEdges(1, 0.0f); // Right edge positions for rows
     for (int si = 0; s[si] != 0; si++) {
         char c = s[si];
         if (s[si] == '\n') {
@@ -355,7 +355,7 @@ void vg::SpriteFont::draw(SpriteBatch* batch, const cString s, const f32v2& posi
     // Get y offset
     f32 yOff = getYOffset(rows.size(), align) * scaling.y;
     // Render each row
-    for (int y = 0; y < rows.size(); y++) {
+    for (size_t y = 0; y < rows.size(); y++) {
         f32 rightEdge = rightEdges[y];
         for (auto& g : rows[y]) {   
             f32v2 position = pos + f32v2(g.x + rightEdges[y] * X_OFF_MULTS[(int)align], yOff + y * m_fontHeight * scaling.y);
@@ -375,25 +375,25 @@ void vg::SpriteFont::checkClipping(const f32v4& clipRect, f32v2& position, f32v2
     if (position.x < clipRect.x) {
         f32 t = clipRect.x - position.x;
         uvRect.x += uvRect.z * (t / size.x);
-        uvRect.z *= 1.0 - (t / size.x);
+        uvRect.z *= 1.0f - (t / size.x);
         position.x = clipRect.x;
         size.x -= t;
     }
     if (position.x + size.x > clipRect.x + clipRect.z) {
         f32 t = position.x + size.x - (clipRect.x + clipRect.z);
-        uvRect.z *= 1.0 - (t / size.x);
+        uvRect.z *= 1.0f - (t / size.x);
         size.x -= t;
     }
     if (position.y < clipRect.y) {
         f32 t = clipRect.y - position.y;
         uvRect.y += uvRect.w * (t / size.y);
-        uvRect.w *= 1.0 - (t / size.y);
+        uvRect.w *= 1.0f - (t / size.y);
         position.y = clipRect.y;
         size.y -= t;
     }
     if (position.y + size.y > clipRect.y + clipRect.w) {
         f32 t = position.y + size.y - (clipRect.y + clipRect.w);
-        uvRect.w *= 1.0 - (t / size.y);
+        uvRect.w *= 1.0f - (t / size.y);
         size.y -= t;
     }
 }
